+++ src/components/movie-page-details/movie-page-details.jsx
@@ -1,0 +1,49 @@
+import React from 'react';
+import {MoviePropTypes} from '../../consts';
+
+const MoviePageDetails = ({movie}) => {
+  const actors = movie.starring.split(`, `).map((actor, i, array) => {
+    return (
+      <React.Fragment key={`actor-${i}-${movie.title}`}>
+        {i < array.length - 1 ? `${actor},` : actor}
+        {(() => i === array.length - 1 ? null : <br/>)()}
+      </React.Fragment>
+    );
+  });
+
+  return (
+    <div className="movie-card__text movie-card__row">
+      <div className="movie-card__text-col">
+        <p className="movie-card__details-item">
+          <strong className="movie-card__details-name">Director</strong>
+          <span className="movie-card__details-value">{movie.director}</span>
+        </p>
+        <p className="movie-card__details-item">
+          <strong className="movie-card__details-name">Starring</strong>
+          <span className="movie-card__details-value">
+            {actors}
+          </span>
+        </p>
+      </div>
+
+      <div className="movie-card__text-col">
+        <p className="movie-card__details-item">
+          <strong className="movie-card__details-name">Run Time</strong>
+          <span className="movie-card__details-value">{movie.runTime}</span>
+        </p>
+        <p className="movie-card__details-item">
+          <strong className="movie-card__details-name">Genre</strong>
+          <span className="movie-card__details-value">{movie.genre}</span>
+        </p>
+        <p className="movie-card__details-item">
+          <strong className="movie-card__details-name">Released</strong>
+          <span className="movie-card__details-value">{movie.year}</span>
+        </p>
+      </div>
+    </div>
+  );
+};
+
+MoviePageDetails.propTypes = MoviePropTypes;
+
+export default MoviePageDetails;
